+++ dk-club-subject/dk-club-infra/src/main/java/com/dk/subject/infra/basic/service/SubjectCategoryService.java
@@ -15,7 +15,7 @@ public interface SubjectCategoryService extends IService<SubjectCategory> {
      * @param id 主键ID
      * @return
      */
-    SubjectCategory getSubjectCategory(Integer id);
+    SubjectCategory getSubjectCategory(Long id);
 
     /**
      * 题目分类获取全部详情
