+++ dk-club-subject/dk-club-infra/src/main/java/com/dk/subject/infra/basic/service/impl/SubjectCategoryServiceImpl.java
@@ -24,7 +24,7 @@ public class SubjectCategoryServiceImpl extends ServiceImpl<SubjectCategoryMappe
     SubjectCategoryMapper subjectCategoryMapper;
 
     @Override
-    public SubjectCategory getSubjectCategory(Integer id) {
+    public SubjectCategory getSubjectCategory(Long id) {
         return subjectCategoryMapper.selectById(id);
     }
